requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-docx>=1.1.0
openai>=1.30.0
anthropic>=0.28.0
